% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_plot.R
\name{river_plot}
\alias{river_plot}
\title{Swiss River Plot Function}
\usage{
river_plot(lakes = T, rivers = T, lakesdetailed = F, riversdetailed = F,
  riverscombined = F, orders = NULL, col_order = F, width_order = F,
  cex_order = 1, col_rhine = "lightblue", col_rhone = "plum",
  col_ticino = "mediumseagreen", col_inn = "salmon", plot_rhine = T,
  plot_rhone = T, plot_ticino = T, plot_inn = T, lines_rhine = T,
  lines_rhone = T, lines_ticino = T, lines_inn = T, border_outline = F,
  width_border = 1, width_country = 1, country_last = T,
  col_country = "black", type_country = 1, step_CH = 5, step_lakes = 10,
  step_rivers = 10, col_water = "blue", axes = c("CH1903", "degree",
  "none"), river_nr = F, scalebar = T)
}
\arguments{
\item{lakes}{Should the lakes be drawn? Defaults to TRUE.}

\item{rivers}{Should the rivers be drawn? Defaults to TRUE.}

\item{lakesdetailed}{Should the detailed version of the lakes be drawn? This option relies on data from swisstopo and requires proper citation if used. Defaults to FALSE.}

\item{riversdetailed}{Should the detailed version of the rivers be drawn? This plots all streams bigger or equals the defined stream order. This option relies on data from swisstopo and requires proper citation if used. Defaults to FALSE.}

\item{riverscombined}{Uses both data sources, a combination of the more detailed swisstopo data and the standard output, to plot the rivers. This option also requires proper citation of the underlying data if used. Defaults to FALSE.}

\item{orders}{The smallest stream order that should be drawn? Defaults to NULL. Highest possible order is 9, lowest is 1.}

\item{col_order}{Colors the rivers according to their stream order. Defaults to FALSE.}

\item{width_order}{Increases river line width with increasing stream order. Defaults to FALSE.}

\item{cex_order}{Controls the width of the river lines. Defined as a scaling factor, where the default value corresponds to 1.}

\item{col_rhine}{Background color of the Rhine drainage area. Defaults to "lightblue".}

\item{col_rhone}{Background color of the Rhone drainage area. Defaults to "plum".}

\item{col_ticino}{Background color of the Ticino drainage area. Defaults to "mediumseagreen".}

\item{col_inn}{Background color of the Inn drainage area. Defaults to "salmon".}

\item{plot_rhine}{Should the Rhine drainage area be plotted as background color? Defaults to TRUE.}

\item{plot_rhone}{Should the Rhone drainage area be plotted as background color? Defaults to TRUE.}

\item{plot_ticino}{Should the Ticino drainage area be plotted as background color? Defaults to TRUE.}

\item{plot_inn}{Should the Inn drainage area be plotted as background color? Defaults to TRUE.}

\item{lines_rhine}{Should the Rhine river lines be plotted? Defaults to TRUE. If set to FALSE, the function relies on the detailed swisstopo data.}

\item{lines_rhone}{Should the Rhone river lines be plotted? Defaults to TRUE. If set to FALSE, the function relies on the detailed swisstopo data.}

\item{lines_ticino}{Should the Ticino river lines be plotted? Defaults to TRUE. If set to FALSE, the function relies on the detailed swisstopo data.}

\item{lines_inn}{Should the Inn river lines be plotted? Defaults to TRUE. If set to FALSE, the function relies on the detailed swisstopo data.}

\item{border_outline}{Only plot drainage area outline instead of filled background. Defaults to FALSE.}

\item{width_border}{Line width of drainage area oulines as a scaling factor. Defaults to 1.}

\item{width_country}{Line width of country border outline as a scaling factor. Defaults to 1.}

\item{country_last}{Country border outline is drawn after catchments. Defaults to TRUE.}

\item{col_country}{Country border outline color. Defaults to "black".}

\item{type_country}{Country border outline linetype. Defaults to 1 (solid).}

\item{step_CH}{Plotting resolution for country border. Increasing the value results in poorer resolution but faster drawing. Defaults to 5.}

\item{step_lakes}{Plotting resolution for lakes. Increasing the value results in poorer resolution but faster drawing. Defaults to 10.}

\item{step_rivers}{Plotting resolution for river lines. Increasing the value results in poorer resolution but faster drawing. Defaults to 10.}

\item{col_water}{Color to be used for waterbodies (lakes and rivers). Defaults to "blue".}

\item{axes}{In what geographic coordinate system should the axis be labelled? Possible options are "CH1903" (Standard Swiss coordinates), "degree" (WGS84), or "none". Defaults to "CH1903"}

\item{river_nr}{Defaults to FALSE.}

\item{scalebar}{Should a scalebar with kilometers be drawn on the plot? Defaults to TRUE.}
}
\description{
This function allows you to produce a figure of the main drainage areas of Switzerland and its lakes, rivers, and streams. The figure appearance is highly customizable, such as defining the stream orders that you want to include in the figure. The four main drainage areas of Switzerland are the Rhine, the Rhone, the Ticino, and the Inn. Parts of the function rely on data from swisstopo. If the function uses them, a prompt will appear to provide the proper citation. Citation of this data is mandatory if used.
}
\examples{
river_plot()
river_plot(border_outline = T, width_border = 2, plot_rhone=F, plot_ticino=F, plot_inn=F, width_order = T, cex_order=0.9, orders=6, lines_rhone=F, lines_ticino=F, lines_inn=F)
}
\keyword{Inn}
\keyword{Rhine}
\keyword{Rhone}
\keyword{Switzerland}
\keyword{Ticino}
\keyword{catchment}
\keyword{drainage}
\keyword{river}
